refactor: split types.ts into types/ directory
Organized by concern:
- types/providers.ts  — LLMProvider type
- types/config.ts     — ToadEyeConfig interface
- types/attributes.ts — GEN_AI_ATTRS, LLM_ATTRS (deprecated)
- types/metrics.ts    — GEN_AI_METRICS, LLM_METRICS (deprecated)
- types/spans.ts      — LLMSpanAttributes, SpanStatus
- types/index.ts      — barrel re-exports

All external imports unchanged (via barrel). Prepares for growth
in batch 4-5 (session types, agent types, budget types).
