+++ packages/instrumentation/src/tracer.ts
@@ -4,7 +4,7 @@ import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-http";
 import { OTLPMetricExporter } from "@opentelemetry/exporter-metrics-otlp-http";
 import { PeriodicExportingMetricReader } from "@opentelemetry/sdk-metrics";
 import { ATTR_SERVICE_NAME } from "@opentelemetry/semantic-conventions";
-import type { ToadEyeConfig } from "./types.js";
+import type { ToadEyeConfig } from "./types/index.js";
 import { initMetrics } from "./metrics.js";
 import { enableAll, disableAll } from "./instrumentations/registry.js";
