+++ packages/instrumentation/src/types/providers.ts
@@ -1,0 +1,5 @@
+/**
+ * LLM providers supported by toad-eye.
+ * Values follow OTel GenAI semconv `gen_ai.provider.name`.
+ */
+export type LLMProvider = "anthropic" | "gemini" | "openai";
